pyproject.toml
[tool.poetry]
name = "calendar"
version = "0.1.0"
description = ""
authors = ["Aleksandr Rozum <rozumalex@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-environ = "^0.4.5"
flake8 = "^3.8.4"
pytest-django = "^4.1.0"
pytest-cov = "^2.10.1"
codecov = "^2.1.10"
psycopg2 = "^2.8.6"
django-timezone-utils = "^0.13"
pytz = "^2020.4"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
